+++ packages/moengage_flutter/moengage_flutter_android/android/src/main/kotlin/com/moengage/flutter/MoEngageFlutterPlugin.kt
@@ -134,6 +134,7 @@ class MoEngageFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
                 METHOD_NAME_SELF_HANDLED_IN_APPS -> getSelfHandledInApps(call, result)
                 METHOD_NAME_IDENTIFY_USER -> identifyUser(call)
                 METHOD_NAME_GET_USER_IDENTITIES -> getUserIdentities(call, result)
+                METHOD_NAME_AUTHENTICATION_DETAILS -> authenticationDetails(call)
                 else ->
                     Logger.print(LogLevel.ERROR) { "$tag onMethodCall() : No mapping for this method." }
             }
@@ -510,6 +511,20 @@ class MoEngageFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
             pluginHelper.identifyUser(context, argument.toString())
         } catch (t: Throwable) {
             Logger.print(LogLevel.ERROR, t) { "$tag identifyUser() : " }
+        }
+    }
+
+    /**
+     *  Passes the JWT authentication details provided in [methodCall] to the native SDK.
+     */
+    private fun authenticationDetails(methodCall: MethodCall) {
+        try {
+            if (methodCall.arguments == null) return
+            val payload = methodCall.arguments.toString()
+            Logger.print { "$tag authenticationDetails() : Arguments: $payload" }
+            pluginHelper.passAuthenticationDetails(context, payload)
+        } catch (t: Throwable) {
+            Logger.print(LogLevel.ERROR, t) { "$tag authenticationDetails() : " }
         }
     }